# apps/api/internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: М3105 API
  version: "1.0"
  summary: Публичное API сайта группы М3105
  description: |
    Публичные эндпоинты сайта группы: предметы, лабораторные с дедлайнами, календарь и ICS-фид,
    конспекты, квизы, ЧаВо, статические страницы, поиск и расписание сдач.

    Все ответы в JSON (кроме `/calendar.ics` и файлов `/files/*`), время в RFC 3339 с таймзоной, даты без времени в формате `YYYY-MM-DD`.
    Возвращаются только опубликованные материалы. Авторизация для этих эндпоинтов не нужна.

    Эндпоинты админки (`/admin/*`) и аккаунтов студентов (`/auth/*`, `/me/*`, запись на сдачи)
    работают по HttpOnly-кукам и описаны в `docs/api.md` репозитория.
  contact:
    url: https://m3105.ru
security: []  # публичные эндпоинты без авторизации
servers:
  - url: https://m3105.ru/api/v1
    description: Продакшен
  - url: http://localhost:8080/api/v1
    description: Локальный API
...
paths:
...
  /comments/{target}/{id}:
    get:
      tags: [social]
      summary: Комментарии к конспекту, лабе или посту
      description: |
        Читать может кто угодно. `POST` с тем же путём и телом `{ body, attachmentIds? }` требует студенческую куку `edu_user`,
        `DELETE /comments/{id}` удаляет свой комментарий. Конспект или лаба должны быть опубликованы.
        `attachmentIds` — id своих файлов из `POST /files`, ещё ни к чему не прикреплённых; с файлами текст можно не писать.
      operationId: listComments
      parameters:
        - name: target
          in: path
          required: true
          schema: { type: string, enum: [note, lab, post] }
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Комментарии, старые первыми
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Comment" }
        "404": { $ref: "#/components/responses/NotFound" }
  /posts:
    get:
      tags: [social]
      summary: Лента постов
      description: |
        Пользовательские посты одной ленты. `POST /posts` с `PostInput` создаёт пост от имени студента,
        `PUT`/`DELETE /posts/{id}` меняют и удаляют свой пост, `PUT`/`DELETE /posts/{id}/like` ставят и снимают лайк.
        Поля `liked` и `mine` считаются для студента, если передана кука `edu_user`.
        Без куки отдаются только посты с `visibility: public`: `members` видят лишь вошедшие студенты,
        и посты с `nsfw: true` всегда `members`. Комментарии к скрытому посту анонимам тоже недоступны (404),
        как и файлы такого поста.
      operationId: listPosts
      parameters:
        - name: kind
          in: query
          required: true
          schema: { type: string, enum: [shawarma, joke] }
        - name: sort
          in: query
          schema: { type: string, enum: [new, top], default: new }
          description: "`top` — сначала посты с наибольшим числом лайков"
      responses:
        "200":
          description: Посты
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Post" }
        "4XX": { $ref: "#/components/responses/ClientError" }
...
components:
...
  responses:
    ClientError:
      description: Некорректный запрос или превышен лимит; формат ошибки одинаковый для всех кодов
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Не найдено или не опубликовано
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Не найдено", code: not_found }
  schemas:
    Error:
      type: object
      description: Формат всех ошибок API
      properties:
        error: { type: string, description: "Сообщение для пользователя" }
        code:
          type: string
          enum: [validation, not_found, conflict, unauthorized, bad_password, rate_limited, internal]
        fields:
          type: object
          additionalProperties: { type: string }
          description: Ошибки валидации по полям (только для `validation`)
      required: [error, code]
    DateTime:
      type: string
      format: date-time
      description: RFC 3339 с таймзоной
...
    Comment:
      type: object
      properties:
        id: { type: integer }
        targetType: { type: string, enum: [note, lab, post] }
        targetId: { type: integer }
        body: { type: string, maxLength: 2000 }
        createdAt: { $ref: "#/components/schemas/DateTime" }
        updatedAt: { $ref: "#/components/schemas/DateTime" }
        authorId: { type: integer }
        authorName: { type: string }
        authorPhotoUrl: { type: string }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        mine: { type: boolean, description: Комментарий текущего студента }
      required: [id, targetType, targetId, body, createdAt, updatedAt, authorId, authorName, authorPhotoUrl, attachments, mine]
    Post:
      type: object
      properties:
        id: { type: integer }
        kind: { type: string, enum: [shawarma, joke] }
        title: { type: string, description: Название точки для шавермы, необязательный заголовок для анекдота }
        body: { type: string, maxLength: 4000 }
        address: { type: string }
        price: { type: [integer, "null"], description: Цена в рублях }
        rating: { type: [integer, "null"], minimum: 1, maximum: 5 }
        visibility: { type: string, enum: [public, members], description: "`members` — только для вошедших студентов" }
        nsfw: { type: boolean, description: Пометка 18+; сайт размывает текст до подтверждения возраста }
        createdAt: { $ref: "#/components/schemas/DateTime" }
        updatedAt: { $ref: "#/components/schemas/DateTime" }
        authorId: { type: integer }
        authorName: { type: string }
        authorPhotoUrl: { type: string }
        likesCount: { type: integer }
        commentsCount: { type: integer }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        liked: { type: boolean }
        mine: { type: boolean }
      required: [id, kind, title, body, address, price, rating, visibility, nsfw, createdAt, updatedAt, authorId, authorName, authorPhotoUrl, likesCount, commentsCount, attachments, liked, mine]
...
    Attachment:
      type: object
      description: Вложение — картинка или файл
      properties:
        id: { type: string }
        url: { type: string, description: "Адрес от корня сайта: /api/v1/files/{id}/{name}" }
        name: { type: string, description: Исходное имя файла }
        contentType: { type: string }
        size: { type: integer, description: Размер в байтах }
        width: { type: [integer, "null"], description: Ширина картинки с учётом поворота из EXIF }
        height: { type: [integer, "null"] }
        createdAt: { $ref: "#/components/schemas/DateTime" }
      required: [id, url, name, contentType, size, width, height, createdAt]